modified:   PhotogateLV.c on branch bugfix-timer
 Functions modified to use Timer1 and not Timer3
 * Time_AllEdges_2Gates()
 * Time_FallingEdges_2Gates()
 * C12_Increment_Counter_on_Timer_Rollover()

The theory is the bug is as a result of Counter being
incremented by both timers. It is not necessary to have
more than one counter with both CPP modules as long as
only one time scale is needed.
The symtom was getting times reported as about twice they
should have been when both CCP's were being used
(like for the collision timer).