Added char buffer and use sprintf()
Added a buffer and used sprintf() to send text to the buffer.
This removes the delays that result from the blocking printf().
This new version is fast enough to catch a sifnificant amount
of switch bounce that occures on some but not all edges.
There are different methods to avoid bounce.