added memory feature to pulse mode
The purpose of this feature is to allow one to retain and display
the first timed duration and not have it overwritten by what comes
later. One must again select Pulse mode with the select button.

The memory feature is turned on by pressing the select button a second
time (the first push on the select button sellects the mode the second
press turns on the memory feature).

If memory is on it will be indicated in the display.